build framework

+++ src/Framework/Database/ORM/Relations/MorphOne.php
@@ -323,7 +323,7 @@ class MorphOne extends Relation
         }
 
         $related = $this->getResults();
-        if ($related === null) {
+        if (!$related instanceof Model) {
             return false;
         }
 
@@ -333,7 +333,12 @@ class MorphOne extends Relation
         }
 
         // Otherwise, perform hard delete
-        return $related->forceDelete();
+        if (method_exists($related, 'forceDelete')) {
+            /** @var Model&\Toporia\Framework\Database\ORM\Concerns\SoftDeletes $related */
+            return $related->forceDelete();
+        }
+
+        return $related->delete();
     }
 
     /**
@@ -370,10 +375,11 @@ class MorphOne extends Relation
             ->whereNotNull($deletedAtColumn)
             ->first();
 
-        if ($related === null) {
+        if (!$related instanceof Model || !method_exists($related, 'restore')) {
             return false;
         }
 
+        /** @var Model&\Toporia\Framework\Database\ORM\Concerns\SoftDeletes $related */
         return $related->restore();
     }
